[CIRC-332] Move location related logic from Item to Location class
